requirements.txt
alembic==1.0.10
aniso8601==6.0.0
argcomplete==1.9.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.159
botocore==1.12.159
certifi==2019.3.9
cffi==1.12.3
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
coverage==4.5.3
cryptography==2.7
docutils==0.14
durationpy==0.5
Faker==1.0.4
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.4.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
future==0.16.0
hjson==3.0.1
idna==2.8
importlib-metadata==0.17
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==3.0.1
kappa==0.6.0
lambda-packages==0.20.0
Mako==1.0.11
MarkupSafe==1.1.1
marshmallow==2.16.0
marshmallow-enum==1.4.1
more-itertools==7.0.0
placebo==0.9.0
pluggy==0.12.0
psycopg2==2.8.2
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.2
pytest==4.4.0
pytest-cov==2.6.1
python-dateutil==2.6.1
python-dotenv==0.10.1
python-editor==1.0.4
python-slugify==1.2.4
pytz==2019.1
PyYAML==5.1
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
SQLAlchemy==1.3.4
text-unidecode==1.2
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.7
Unidecode==1.0.23
urllib3==1.25.3
Werkzeug==0.15.4
wsgi-request-logger==0.4.6
zappa==0.48.2
zipp==0.5.1
